[INPL-6150] Create connection to Shipment domain and DeliveryProtection module

+++ RouteTrackerFlow/RouteTrackerFlowCore/src/main/java/com/warehouse/routetracker/infrastructure/adapter/secondary/mapper/RouteModelMapper.java
@@ -77,7 +77,7 @@ public interface RouteModelMapper {
         final Long userId = route.getUserId();
         if (Objects.nonNull(userId)) {
             return UserEntity.builder()
-                    .id(userId.intValue())
+                    .id(userId)
                     .build();
         }
         return null;
